requirements.txt
# ── Core ML & LLM ─────────────────────────────────────────────────────────────
torch>=2.1.0
transformers>=4.38.0
peft>=0.9.0
trl>=0.8.0
accelerate>=0.27.0
datasets>=2.17.0
sentence-transformers>=2.6.0

# ── Vector Search ─────────────────────────────────────────────────────────────
faiss-cpu>=1.7.4

# ── Evaluation ────────────────────────────────────────────────────────────────
scikit-learn>=1.4.0
nltk>=3.8.1
rouge-score>=0.1.2
bert-score>=0.3.13
detoxify>=0.5.2

# ── Data & Numerics ───────────────────────────────────────────────────────────
numpy>=1.26.0
pandas>=2.2.0

# ── Visualisation ─────────────────────────────────────────────────────────────
matplotlib>=3.8.0
seaborn>=0.13.0

# ── UI ────────────────────────────────────────────────────────────────────────
gradio>=4.19.0

# ── Utilities ─────────────────────────────────────────────────────────────────
tqdm>=4.66.0
sacremoses>=0.1.1
sentencepiece>=0.2.0
